FIXED LAYOUT
put classes in respective dart files
